Refactor template generation
